Feat : 상영 중인 영화들 반환 api 구현 완료
infra 모듈에 필요한 mapper 및 adapter 구현
application 모듈에서 MovieService에서 dto의 List 만들어 반환

+++ module_infrastructure/src/main/java/project/redis/movie/repository/MovieRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import project.redis.movie.entity.MovieEntity;
 
 @Repository
-public interface MovieRepository extends JpaRepository<MovieEntity, Integer> {
+public interface MovieRepository extends JpaRepository<MovieEntity, Long> {
     List<MovieEntity> findAll();
 }
